ANDROID: sched/walt: Fix lockdep warning
When enabling lockdep on bluecross, a warning fires on boot up in the
WALT code. This is because the walt_irq_work handler acquires rq-locks
in succession however, this is forbidden. To fix the warning, we use the
raw_spin_lock_nested API to tell lockdep we are intentionally acquiring
the rq-lock in a nested fashion.

Bug: 110360156
Change-Id: I8598d79632991d799edcc8808d2e2f383b7a7ad3
Signed-off-by: Joel Fernandes (Google) <joel@joelfernandes.org>